iNextLabs/Projects/Task 2/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
requests
python-dotenv
pydantic